remove link utils provider
