+++ src/components/FormatFilter.tsx
@@ -3,9 +3,9 @@ import { jsx } from "theme-ui";
 import * as React from "react";
 import FormLabel from "./form/FormLabel";
 import Router from "next/router";
-import useLinkUtils from "./context/LinkUtilsContext";
 import Select from "./Select";
 import { CollectionData } from "interfaces";
+import useLinkUtils from "hooks/useLinkUtils";
 
 /**
  * This filter depends on the "Formats" facetGroup, which should have
