+++ src/components/bookDetails/FulfillmentCard.tsx
@@ -39,7 +39,7 @@ import useReadOnlineButton from "hooks/useReadOnlineButton";
 import { APP_CONFIG } from "config";
 import track from "analytics/track";
 import { useRouter } from "next/router";
-import useLinkUtils from "components/context/LinkUtilsContext";
+import useLinkUtils from "hooks/useLinkUtils";
 
 const FulfillmentCard: React.FC<{ book: AnyBook }> = ({ book }) => {
   return (
