+++ src/components/context/ContextProvider.tsx
@@ -4,7 +4,6 @@ import { LibraryProvider } from "./LibraryContext";
 import { Provider as ReakitProvider } from "reakit";
 import { ThemeProvider } from "theme-ui";
 import makeTheme from "../../theme";
-import { LinkUtilsProvider } from "./LinkUtilsContext";
 import { UserProvider } from "components/context/UserContext";
 import AuthModal from "auth/AuthModal";
 import { ConfigInterface, SWRConfig } from "swr";
@@ -30,11 +29,9 @@ const AppContextProvider: React.FC<ProviderProps> = ({ children, library }) => {
       <ThemeProvider theme={theme}>
         <ReakitProvider>
           <LibraryProvider library={library}>
-            <LinkUtilsProvider library={library}>
-              <UserProvider>
-                <AuthModal>{children}</AuthModal>
-              </UserProvider>
-            </LinkUtilsProvider>
+            <UserProvider>
+              <AuthModal>{children}</AuthModal>
+            </UserProvider>
           </LibraryProvider>
         </ReakitProvider>
       </ThemeProvider>
